+++ ThirdPartyAdapters/unity/unity/src/test/kotlin/com/google/ads/mediation/unity/UnityRewardedAdTest.kt
@@ -4,7 +4,9 @@ import android.app.Activity
 import androidx.core.os.bundleOf
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.ads.mediation.unity.UnityAdsAdapterUtils.getMediationErrorCode
+import com.google.ads.mediation.unity.UnityAdsAdapterUtils.WATERMARK
+import com.google.ads.mediation.unity.UnityAdsAdapterUtils.getMediationLoadErrorCode
+import com.google.ads.mediation.unity.UnityAdsAdapterUtils.getMediationShowErrorCode
 import com.google.ads.mediation.unity.UnityMediationAdapter.ADAPTER_ERROR_DOMAIN
 import com.google.ads.mediation.unity.UnityMediationAdapter.ERROR_CONTEXT_NOT_ACTIVITY
 import com.google.ads.mediation.unity.UnityMediationAdapter.ERROR_MSG_NON_ACTIVITY
@@ -15,12 +17,12 @@ import com.google.android.gms.ads.mediation.MediationRewardedAd
 import com.google.android.gms.ads.mediation.MediationRewardedAdCallback
 import com.google.android.gms.ads.mediation.MediationRewardedAdConfiguration
 import com.google.common.truth.Truth.assertThat
-import com.unity3d.ads.IUnityAdsInitializationListener
-import com.unity3d.ads.UnityAds
+import com.unity3d.ads.InitializationListener
+import com.unity3d.ads.RewardedAd
+import com.unity3d.ads.ShowConfiguration
+import com.unity3d.ads.ShowFinishState
 import com.unity3d.ads.UnityAds.UnityAdsLoadError
-import com.unity3d.ads.UnityAds.UnityAdsShowError
-import com.unity3d.ads.UnityAdsLoadOptions
-import com.unity3d.ads.UnityAdsShowOptions
+import com.unity3d.ads.UnityAdsError
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -29,12 +31,11 @@ import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.doAnswer
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.eq
-import org.mockito.kotlin.isNull
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
-import org.mockito.kotlin.notNull
 import org.mockito.kotlin.spy
 import org.mockito.kotlin.verify
+import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.whenever
 import org.robolectric.Robolectric
 
@@ -70,7 +71,7 @@ class UnityRewardedAdTest {
 
   @Test
   fun onUnityAdsAdLoaded_invokesOnSuccess() {
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+    unityRewardedAd.loadListener.onAdLoaded(mock(), null)
 
     verify(rewardedAdLoadCallback).onSuccess(unityRewardedAd)
   }
@@ -78,14 +79,12 @@ class UnityRewardedAdTest {
   @Test
   fun onUnityAdsFailedToLoad_invokesOnFailure() {
     val errorCaptor = argumentCaptor<AdError>()
-    val unityAdsLoadError = UnityAdsLoadError.NO_FILL
-    val errorCode = getMediationErrorCode(unityAdsLoadError)
+    val unityAdsLoadError = mock<UnityAdsError>()
+    whenever(unityAdsLoadError.code).doReturn(52100)
+    whenever(unityAdsLoadError.message).doReturn(TEST_ERROR_MESSAGE)
+    val errorCode = getMediationLoadErrorCode(unityAdsLoadError)
 
-    unityRewardedAd.unityLoadListener.onUnityAdsFailedToLoad(
-      TEST_PLACEMENT_ID,
-      unityAdsLoadError,
-      TEST_ERROR_MESSAGE,
-    )
+    unityRewardedAd.loadListener.onAdLoaded(null, unityAdsLoadError)
 
     verify(rewardedAdLoadCallback).onFailure(errorCaptor.capture())
     val capturedError = errorCaptor.firstValue
@@ -96,9 +95,10 @@ class UnityRewardedAdTest {
 
   @Test
   fun onUnityAdsShowStart_invokesOnAdOpenedReportAdImpressionAndOnVideoStart() {
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+    val rewardedAd = mock<RewardedAd>()
+    unityRewardedAd.loadListener.onAdLoaded(rewardedAd, null)
 
-    unityRewardedAd.unityShowListener.onUnityAdsShowStart(TEST_PLACEMENT_ID)
+    unityRewardedAd.unityShowListener.onStarted(rewardedAd)
 
     verify(rewardedAdCallback).onAdOpened()
     verify(rewardedAdCallback).reportAdImpression()
@@ -107,7 +107,9 @@ class UnityRewardedAdTest {
 
   @Test
   fun onUnityAdsShowStart_withNullAdCallback_doesNotInvokeAnyCallbackMethod() {
-    unityRewardedAd.unityShowListener.onUnityAdsShowStart(TEST_PLACEMENT_ID)
+    val rewardedAd = mock<RewardedAd>()
+
+    unityRewardedAd.unityShowListener.onStarted(rewardedAd)
 
     verify(rewardedAdCallback, never()).onAdOpened()
     verify(rewardedAdCallback, never()).reportAdImpression()
@@ -116,41 +118,45 @@ class UnityRewardedAdTest {
 
   @Test
   fun onUnityAdsShowClick_invokesReportAdClicked() {
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+    val rewardedAd = mock<RewardedAd>()
+    unityRewardedAd.loadListener.onAdLoaded(rewardedAd, null)
 
-    unityRewardedAd.unityShowListener.onUnityAdsShowClick(TEST_PLACEMENT_ID)
+    unityRewardedAd.unityShowListener.onClicked(rewardedAd)
 
     verify(rewardedAdCallback).reportAdClicked()
   }
 
   @Test
   fun onUnityAdsShowClick_withNullAdCallback_doesNotInvokeReportAdClicked() {
-    unityRewardedAd.unityShowListener.onUnityAdsShowClick(TEST_PLACEMENT_ID)
+    unityRewardedAd.unityShowListener.onClicked(mock())
 
     verify(rewardedAdCallback, never()).reportAdClicked()
   }
 
   @Test
-  fun onUnityAdsShowComplete_withStateCompleted_invokesOnVideoCompleteOnUserEarnedRewardAndOnAdClosed() {
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+  fun onUnityAdsShowComplete_withStateCompleted_invokesOnVideoCompleteAndOnAdClosed() {
+    val rewardedAd = mock<RewardedAd>()
+    unityRewardedAd.loadListener.onAdLoaded(rewardedAd, null)
 
-    unityRewardedAd.unityShowListener.onUnityAdsShowComplete(
-      TEST_PLACEMENT_ID,
-      UnityAds.UnityAdsShowCompletionState.COMPLETED,
+    unityRewardedAd.unityShowListener.onCompleted(
+      rewardedAd,
+      ShowFinishState.COMPLETED,
     )
 
     verify(rewardedAdCallback).onVideoComplete()
-    verify(rewardedAdCallback).onUserEarnedReward()
     verify(rewardedAdCallback).onAdClosed()
+
+    // this is triggered by a new callback. Should not be triggered here
+    verify(rewardedAdCallback, never()).onUserEarnedReward()
   }
 
   @Test
   fun onUnityAdsShowComplete_withStateNotCompleted_invokesOnlyOnAdClosed() {
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+    unityRewardedAd.loadListener.onAdLoaded(mock(), null)
 
-    unityRewardedAd.unityShowListener.onUnityAdsShowComplete(
-      TEST_PLACEMENT_ID,
-      UnityAds.UnityAdsShowCompletionState.SKIPPED,
+    unityRewardedAd.unityShowListener.onCompleted(
+      mock(),
+      ShowFinishState.SKIPPED,
     )
 
     verify(rewardedAdCallback, never()).onVideoComplete()
@@ -160,28 +166,25 @@ class UnityRewardedAdTest {
 
   @Test
   fun onUnityAdsShowComplete_withNullAdCallback_doesNotInvokeAnyCallbackMethod() {
-    unityRewardedAd.unityShowListener.onUnityAdsShowComplete(
-      TEST_PLACEMENT_ID,
-      UnityAds.UnityAdsShowCompletionState.COMPLETED,
+    unityRewardedAd.unityShowListener.onCompleted(
+      mock(),
+      ShowFinishState.COMPLETED,
     )
 
     verify(rewardedAdCallback, never()).onVideoComplete()
-    verify(rewardedAdCallback, never()).onUserEarnedReward()
     verify(rewardedAdCallback, never()).onAdClosed()
   }
 
   @Test
   fun onUnityAdsShowFailure_invokesOnAdFailedToShow() {
     val errorCaptor = argumentCaptor<AdError>()
-    val unityAdsShowError = UnityAdsShowError.INTERNAL_ERROR
-    val errorCode = getMediationErrorCode(unityAdsShowError)
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+    val unityAdsError = mock<UnityAdsError>()
+    whenever(unityAdsError.message).doReturn(TEST_ERROR_MESSAGE)
+    whenever(unityAdsError.code).doReturn(52202)
+    val errorCode = getMediationShowErrorCode(unityAdsError)
+    unityRewardedAd.loadListener.onAdLoaded(mock(), null)
 
-    unityRewardedAd.unityShowListener.onUnityAdsShowFailure(
-      TEST_PLACEMENT_ID,
-      unityAdsShowError,
-      TEST_ERROR_MESSAGE,
-    )
+    unityRewardedAd.unityShowListener.onFailed(mock(), unityAdsError)
 
     verify(rewardedAdCallback).onAdFailedToShow(errorCaptor.capture())
     val capturedError = errorCaptor.firstValue
@@ -192,18 +195,14 @@ class UnityRewardedAdTest {
 
   @Test
   fun onUnityAdsShowFailure_withNullAdCallback_doesNotInvokeOnAdFailedToShow() {
-    unityRewardedAd.unityShowListener.onUnityAdsShowFailure(
-      TEST_PLACEMENT_ID,
-      UnityAdsShowError.INTERNAL_ERROR,
-      TEST_ERROR_MESSAGE,
-    )
+    unityRewardedAd.unityShowListener.onFailed(mock(), mock())
 
     verify(rewardedAdCallback, never()).onAdFailedToShow(any<AdError>())
   }
 
   @Test
   fun showAd_withNonActivityContext_invokesOnAdFailedToShow() {
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_PLACEMENT_ID)
+    unityRewardedAd.loadListener.onAdLoaded(mock(), null)
     val errorCaptor = argumentCaptor<AdError>()
 
     unityRewardedAd.showAd(ApplicationProvider.getApplicationContext())
@@ -219,60 +218,57 @@ class UnityRewardedAdTest {
   fun showAd_invokesUnityAdsShow() {
     doAnswer { invocation ->
         val args = invocation.arguments
-        (args[2] as IUnityAdsInitializationListener).onInitializationComplete()
+        (args[1] as InitializationListener).onInitializationComplete(null)
       }
       .whenever(unityInitializer)
-      .initializeUnityAds(any(), any(), any())
+      .initializeUnityAds(any(), any())
     whenever(rewardedAdConfiguration.serverParameters) doReturn
       bundleOf(
         UnityMediationAdapter.KEY_PLACEMENT_ID to TEST_PLACEMENT_ID,
         UnityMediationAdapter.KEY_GAME_ID to TEST_GAME_ID,
       )
     whenever(rewardedAdConfiguration.context) doReturn activity
-    val unityAdsLoadOptions: UnityAdsLoadOptions = mock()
-    val unityAdsShowOptions: UnityAdsShowOptions = mock()
-    whenever(unityAdsLoader.createUnityAdsLoadOptionsWithId(any())) doReturn unityAdsLoadOptions
-    whenever(unityAdsLoader.createUnityAdsShowOptionsWithId(any())) doReturn unityAdsShowOptions
     unityRewardedAd.loadAd(rewardedAdConfiguration)
-    unityRewardedAd.unityLoadListener.onUnityAdsAdLoaded(TEST_LOADED_PLACEMENT_ID)
+
+    val loadedAd = mock<RewardedAd>()
+
+    unityRewardedAd.loadListener.onAdLoaded(loadedAd, null)
 
     unityRewardedAd.showAd(activity)
 
-    verify(unityAdsLoader).createUnityAdsShowOptionsWithId(notNull())
-    verify(unityAdsShowOptions).set(UnityMediationAdapter.KEY_WATERMARK, TEST_WATERMARK)
-    verify(unityAdsLoader).show(any(), eq(TEST_LOADED_PLACEMENT_ID), any(), any())
+    val configCaptor = argumentCaptor<ShowConfiguration>()
+    verify(loadedAd).show(eq(activity), configCaptor.capture(), any())
+    assertThat(configCaptor.firstValue.extras[WATERMARK]).isEqualTo(TEST_WATERMARK)
   }
 
   @Test
   fun showAd_withNullPlacementId_invokesUnityAdsShowWithNullId() {
     doAnswer { invocation ->
         val args = invocation.arguments
-        (args[2] as IUnityAdsInitializationListener).onInitializationComplete()
+        (args[1] as InitializationListener).onInitializationComplete(null)
       }
       .whenever(unityInitializer)
-      .initializeUnityAds(any(), any(), any())
+      .initializeUnityAds(any(), any())
     whenever(rewardedAdConfiguration.serverParameters) doReturn
       bundleOf(
         UnityMediationAdapter.KEY_PLACEMENT_ID to TEST_PLACEMENT_ID,
         UnityMediationAdapter.KEY_GAME_ID to TEST_GAME_ID,
       )
     whenever(rewardedAdConfiguration.context) doReturn activity
-    val unityAdsLoadOptions: UnityAdsLoadOptions = mock()
-    val unityAdsShowOptions: UnityAdsShowOptions = mock()
-    whenever(unityAdsLoader.createUnityAdsLoadOptionsWithId(any())) doReturn unityAdsLoadOptions
-    whenever(unityAdsLoader.createUnityAdsShowOptionsWithId(any())) doReturn unityAdsShowOptions
     unityRewardedAd.loadAd(rewardedAdConfiguration)
     val unityAdsLoadError = UnityAdsLoadError.NO_FILL
-    unityRewardedAd.unityLoadListener.onUnityAdsFailedToLoad(
+
+    unityRewardedAd.loadListener.onAdLoaded(
       null,
-      unityAdsLoadError,
-      TEST_ERROR_MESSAGE,
+      mock<UnityAdsError>().apply {
+        whenever(this.code).doReturn(0)
+        whenever(this.message).doReturn(TEST_ERROR_MESSAGE)
+      }
     )
 
     unityRewardedAd.showAd(activity)
 
-    verify(unityAdsLoader).createUnityAdsShowOptionsWithId(notNull())
-    verify(unityAdsLoader).show(any(), isNull(), any(), any())
+    verifyNoInteractions(rewardedAdCallback)
   }
 
   companion object {